third_party/blink/web_tests/platform/generic/transitions/unprefixed-transform-origin-expected.txt

all 10ms
fired transitionend with propertyName: transform-origin


transform-origin 10ms
fired transitionend with propertyName: transform-origin


-webkit-transform-origin 10ms
fired transitionend with propertyName: transform-origin


-webkit-transform-origin-x 10ms
fired transitionend with propertyName: -webkit-transform-origin-x


-webkit-transform-origin-y 10ms
fired transitionend with propertyName: -webkit-transform-origin-y


-webkit-transform-origin-z 10ms
fired transitionend with propertyName: -webkit-transform-origin-z
